Add second constraint to RichRope
This was the product of a conversation on the Discourse:
https://discourse.julialang.org/t/why-are-self-referential-struct-parameters-impossible/110982

This has two clear advantages: leaf types are smaller, and it
allows factoring of methods by leaf and branch variants.  The
impact on type stability is TBD.

Regardless it will serve for a reference implementation, and,
perhaps, a motivating example.