All tests green
Going two indices back instead of just one feels... unprincipled,
somehow, like I'm going to have to wrestle with this thing again
when I start applying Zalgo text to the test suite. That said,
it works with Farmer Bob, so it's pretty close to correct.

I'll need to include some notes in the documentation about how to
preserve the grapheme invariant, this doesn't appear to be well
understood. Pointless until I add the grapheme-using variations
of the basic interface, which is itself nearly done. So that's
nice.