Graphemes
Needed to break off the necessary logic into something testable,
now this implementation is robust, or appears to be, and is at
least rounding the corner on optimal.  Need to add StringViews to
get copies down to the bare necessity, and maybe if the compiler
can prove that nothing uses the write vector before it gets dunked
into the remainder it won't need to double-allocate for it.