AbstractTree tests
I've taken leafsize of 1 off the table, because there's an alarming
bug where the type checks in concatenation fail.  Alarming because
the type it says is a fail is the same type, as far as I can
determine, as the type it evaluates to.  Really seems like a Julia
bug but everything which actually uses one-Char leaves is behaving
itself at 1, so ¯\_(ツ)_/¯