Add rebuild(rope)
This isn't called "balance", which I'm reserving as the name of a
function which puts some effort into preserving parts of the rope
which don't need attention.  I don't like the heuristics I've seen
for detecting imbalance and am not currently up to writing and
proving anything of the sort.