Somewhat futile attempt at type stability
It helps a bit to make the Vector a Union but that AbstractRope
kludge fucks me sideways on stability. If I could get the type
system to encode the actual type of the children, but no further,
that would link the chain. I have to keep "reminding" the runtime
that, no, /by construction/, there is no other AbstractRope{S}
possible. Meh.