Use cursor for general iteration
For reasons I don't entirely understand, this is much faster and
uses half the memory. Possible this is due to more-complete type
annotation of the cursor, but I'll take it: using `collect` this
is only five times slower than iterating the same data as a String,
and it's unlikely that I can improve on that.