Add _nextleaf (probably removing it)
Just committing this in case I need it again.

Slowly realizing that the data structure I'm looking for here is
a lazy zipper.  Instead of these ridiculous stacks, every ZipRope
holds a reference to its parent, and whether it's a left child or
a right child.

This is probably reimplementing code from AbstractTrees but I don't
care about that at this point, I just want things to Work.