Add RopeZip
These are stateless, and can be run in both directions.

A bit more allocation but (probably, have to test) more type stable.

Ok, a lot more allocation, we've gone from log(n) to n, but these
are a lot easier to work with and I don't need to ditch the existing
stateful iterators just because I have this.