Suspiciously fast reverse iteration
All of @benchmark, @btime, and @time, claim that this is absurdly
faster than reverse-iterating an equivalent String, or, forward
iterating the same rope, or, forward iterating the same String.

I find this implausible on first principles. But the benchmarks
won't shut up about it and I find that distressing.

Call it acceptably fast...