Corrects bug in ReversedRope
This was an interesting bug to catch, because I was testing the
iterator with benchmarks against the string-equivalent of the rope,
and Iterators.reverse(rope) was way too fast, microseconds fast.

And when I'd compare the results of collect between the string and
the rope (oops), it was the same!

This is because the garbage collector kept freeing those vectors,
then the allocator found a perfect-sized block of memory in the
orphanage. The first iteration replaced the first element of that
block with the correct value, the second iteration returned nothing,
and so the rest of the vector was "garbage" which just happened to
contain the exact data I was testing for.

Funny stuff. This version should work correctly, and it's only
~1.5X slower than reverse-iterating the actual rope, so that's
enough for me.